+++ packages/liveblocks-core/src/__tests__/room.test.ts
@@ -3,10 +3,11 @@ import { setupServer } from "msw/node";
 
 import type { RoomAuthToken } from "../AuthToken";
 import { RoomScope } from "../AuthToken";
+import { LiveList } from "../crdts/LiveList";
+import type { LsonObject } from "../crdts/Lson";
 import { lsonToJson } from "../immutable";
 import * as console from "../lib/fancy-console";
 import type { Json, JsonObject } from "../lib/Json";
-import { LiveList } from "../LiveList";
 import type { BaseUserMeta } from "../protocol/BaseUserMeta";
 import { ClientMsgCode } from "../protocol/ClientMsg";
 import type { IdTuple, SerializedCrdt } from "../protocol/SerializedCrdt";
@@ -18,7 +19,6 @@ import {
   _private_makeStateMachine as makeStateMachine,
   createRoom,
 } from "../room";
-import type { LsonObject } from "../types/Lson";
 import type { Others } from "../types/Others";
 import { WebsocketCloseCodes } from "../types/WebsocketCloseCodes";
 import {
