+++ packages/liveblocks-core/src/crdts/AbstractCrdt.ts
@@ -1,10 +1,10 @@
-import { assertNever } from "./lib/assert";
-import type { CreateChildOp, Op } from "./protocol/Op";
-import { OpCode } from "./protocol/Op";
-import type { SerializedCrdt } from "./protocol/SerializedCrdt";
-import type { Immutable } from "./types/Immutable";
-import type { LiveNode } from "./types/Lson";
-import type { StorageUpdate } from "./types/StorageUpdates";
+import { assertNever } from "../lib/assert";
+import type { CreateChildOp, Op } from "../protocol/Op";
+import { OpCode } from "../protocol/Op";
+import type { SerializedCrdt } from "../protocol/SerializedCrdt";
+import type { Immutable } from "../types/Immutable";
+import type { StorageUpdate } from "../types/StorageUpdates";
+import type { LiveNode } from "./Lson";
 
 export type ApplyResult =
   | { reverse: Op[]; modified: StorageUpdate }
