+++ packages/liveblocks-core/src/crdts/LiveMap.ts
@@ -1,22 +1,22 @@
-import type { ApplyResult, ManagedPool } from "./AbstractCrdt";
-import { AbstractCrdt, OpSource } from "./AbstractCrdt";
-import { nn } from "./lib/assert";
-import { freeze } from "./lib/freeze";
+import { nn } from "../lib/assert";
+import { freeze } from "../lib/freeze";
 import {
   creationOpToLiveNode,
   deserialize,
   isLiveNode,
   liveNodeToLson,
   lsonToLiveNode,
-} from "./liveblocks-helpers";
-import type { CreateChildOp, CreateMapOp, Op } from "./protocol/Op";
-import { OpCode } from "./protocol/Op";
-import type { IdTuple, SerializedMap } from "./protocol/SerializedCrdt";
-import { CrdtType } from "./protocol/SerializedCrdt";
-import type { LiveNode, Lson } from "./types/Lson";
-import type { ParentToChildNodeMap } from "./types/NodeMap";
-import type { LiveMapUpdates } from "./types/StorageUpdates";
-import type { ToImmutable } from "./types/ToImmutable";
+} from "../liveblocks-helpers";
+import type { CreateChildOp, CreateMapOp, Op } from "../protocol/Op";
+import { OpCode } from "../protocol/Op";
+import type { IdTuple, SerializedMap } from "../protocol/SerializedCrdt";
+import { CrdtType } from "../protocol/SerializedCrdt";
+import type { ParentToChildNodeMap } from "../types/NodeMap";
+import type { LiveMapUpdates } from "../types/StorageUpdates";
+import type { ToImmutable } from "../types/ToImmutable";
+import type { ApplyResult, ManagedPool } from "./AbstractCrdt";
+import { AbstractCrdt, OpSource } from "./AbstractCrdt";
+import type { LiveNode, Lson } from "./Lson";
 
 /**
  * The LiveMap class is similar to a JavaScript Map that is synchronized on all clients.
