+++ packages/liveblocks-core/src/crdts/LiveObject.ts
@@ -1,14 +1,13 @@
-import type { ApplyResult, ManagedPool } from "./AbstractCrdt";
-import { AbstractCrdt, OpSource } from "./AbstractCrdt";
-import { nn } from "./lib/assert";
-import type { JsonObject } from "./lib/Json";
-import { fromEntries } from "./lib/utils";
+import type { LiveNode, Lson, LsonObject } from "../crdts/Lson";
+import { nn } from "../lib/assert";
+import type { JsonObject } from "../lib/Json";
+import { fromEntries } from "../lib/utils";
 import {
   creationOpToLson,
   deserializeToLson,
   isLiveNode,
   isLiveStructure,
-} from "./liveblocks-helpers";
+} from "../liveblocks-helpers";
 import type {
   CreateChildOp,
   CreateObjectOp,
@@ -17,22 +16,23 @@ import type {
   DeleteObjectKeyOp,
   Op,
   UpdateObjectOp,
-} from "./protocol/Op";
-import { OpCode } from "./protocol/Op";
+} from "../protocol/Op";
+import { OpCode } from "../protocol/Op";
 import type {
   IdTuple,
   SerializedObject,
   SerializedRootObject,
-} from "./protocol/SerializedCrdt";
-import { CrdtType } from "./protocol/SerializedCrdt";
-import type { LiveNode, Lson, LsonObject } from "./types/Lson";
-import type { ParentToChildNodeMap } from "./types/NodeMap";
+} from "../protocol/SerializedCrdt";
+import { CrdtType } from "../protocol/SerializedCrdt";
+import type { ParentToChildNodeMap } from "../types/NodeMap";
 import type {
   LiveObjectUpdateDelta,
   LiveObjectUpdates,
   UpdateDelta,
-} from "./types/StorageUpdates";
-import type { ToImmutable } from "./types/ToImmutable";
+} from "../types/StorageUpdates";
+import type { ToImmutable } from "../types/ToImmutable";
+import type { ApplyResult, ManagedPool } from "./AbstractCrdt";
+import { AbstractCrdt, OpSource } from "./AbstractCrdt";
 
 /**
  * The LiveObject class is similar to a JavaScript object that is synchronized on all clients.
