+++ packages/liveblocks-core/src/crdts/__tests__/AuthToken.test.ts
@@ -1,5 +1,5 @@
-import type { JwtMetadata } from "../AuthToken";
-import { isTokenExpired, parseRoomAuthToken, RoomScope } from "../AuthToken";
+import type { JwtMetadata } from "../../AuthToken";
+import { isTokenExpired, parseRoomAuthToken, RoomScope } from "../../AuthToken";
 
 describe("isTokenExpired", () => {
   const MINUTES = 60 * 1000;
