+++ packages/liveblocks-core/src/immutable.ts
@@ -1,3 +1,8 @@
+import { LiveList } from "./crdts/LiveList";
+import { LiveMap } from "./crdts/LiveMap";
+import { LiveObject } from "./crdts/LiveObject";
+import { LiveRegister } from "./crdts/LiveRegister";
+import type { LiveNode, Lson, LsonObject, ToJson } from "./crdts/Lson";
 import * as console from "./lib/fancy-console";
 import type { Json, JsonObject } from "./lib/Json";
 import { isPlainObject } from "./lib/utils";
@@ -6,11 +11,6 @@ import {
   isLiveList,
   isLiveObject,
 } from "./liveblocks-helpers";
-import { LiveList } from "./LiveList";
-import { LiveMap } from "./LiveMap";
-import { LiveObject } from "./LiveObject";
-import { LiveRegister } from "./LiveRegister";
-import type { LiveNode, Lson, LsonObject, ToJson } from "./types/Lson";
 import type { StorageUpdate } from "./types/StorageUpdates";
 
 function lsonObjectToJson<O extends LsonObject>(
